# perf: support JSON as Registry Configuration