--- a/pkg/install-registry/install.go
+++ b/pkg/install-registry/install.go
@@ -2,6 +2,7 @@ package registry
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"os"
 	"path/filepath"
@@ -67,6 +68,24 @@ func (inst *installer) InstallRegistries(ctx context.Context, cfg *config.Config
 	return registryContents, nil
 }
 
+func readRegistry(p string, registry *config.RegistryContent) error {
+	f, err := os.Open(p)
+	if err != nil {
+		return fmt.Errorf("open the registry configuration file: %w", err)
+	}
+	defer f.Close()
+	if filepath.Ext(p) == ".json" {
+		if err := json.NewDecoder(f).Decode(registry); err != nil {
+			return fmt.Errorf("parse the registry configuration as JSON: %w", err)
+		}
+		return nil
+	}
+	if err := yaml.NewDecoder(f).Decode(registry); err != nil {
+		return fmt.Errorf("parse the registry configuration as YAML: %w", err)
+	}
+	return nil
+}
+
 // installRegistry installs and reads the registry file and returns the registry content.
 // If the registry file already exists, the installation is skipped.
 func (inst *installer) installRegistry(ctx context.Context, registry *config.Registry, cfgFilePath string, logE *logrus.Entry) (*config.RegistryContent, error) {
@@ -79,14 +98,9 @@ func (inst *installer) installRegistry(ctx context.Context, registry *config.Reg
 		return inst.getRegistry(ctx, registry, registryFilePath, logE)
 	}
 
-	f, err := os.Open(registryFilePath)
-	if err != nil {
-		return nil, fmt.Errorf("open the registry configuration file: %w", err)
-	}
-	defer f.Close()
 	registryContent := &config.RegistryContent{}
-	if err := yaml.NewDecoder(f).Decode(registryContent); err != nil {
-		return nil, fmt.Errorf("parse the registry configuration: %w", err)
+	if err := readRegistry(registryFilePath, registryContent); err != nil {
+		return nil, err
 	}
 	return registryContent, nil
 }
@@ -114,8 +128,14 @@ func (inst *installer) getGitHubContentRegistry(ctx context.Context, registry *c
 		return nil, fmt.Errorf("write the configuration file: %w", err)
 	}
 	registryContent := &config.RegistryContent{}
+	if filepath.Ext(registryFilePath) == ".json" {
+		if err := json.Unmarshal(b, registryContent); err != nil {
+			return nil, fmt.Errorf("parse the registry configuration file as JSON: %w", err)
+		}
+		return registryContent, nil
+	}
 	if err := yaml.Unmarshal(b, registryContent); err != nil {
-		return nil, fmt.Errorf("parse the registry configuration file: %w", err)
+		return nil, fmt.Errorf("parse the registry configuration file as YAML: %w", err)
 	}
 	return registryContent, nil
 }