Close #677

## Benchmark Tests

Compare JSON with YAML.

```console
$ go test -bench . -benchmem
goos: darwin
goarch: arm64
pkg: github.com/aquaproj/aqua/pkg/config
BenchmarkReadRegistry/yaml-8         	     198	   5812580 ns/op	 2613083 B/op	   59282 allocs/op
BenchmarkReadRegistry/json-8         	     588	   1711825 ns/op	 1087863 B/op	   10940 allocs/op
PASS
ok  	github.com/aquaproj/aqua/pkg/config	5.779s
```

JSON is much faster than YAML.

## Benchmark Tests with hyperfine

https://github.com/sharkdp/hyperfine

```console
$ hyperfine --version
hyperfine 1.13.0
```

```yaml
---
# aqua - Declarative CLI Version Manager
# https://aquaproj.github.io/
registries:
  - type: local
    name: local
    path: registry.json # registry.yaml

packages:
  - name: suzuki-shunsuke/tfcmt@v3.2.1
    registry: local
```

```console
$ hyperfine -N --warmup 3 'tfcmt -v' '/Users/shunsukesuzuki/.local/share/aquaproj-aqua/pkgs/github_release/github.com/suzuki-shunsuke/tfcmt/v3.2.1/tfcmt_darwin_arm64.tar.gz/tfcmt -v'
Benchmark 1: tfcmt -v
  Time (mean ± σ):      19.2 ms ±   0.3 ms    [User: 15.3 ms, System: 5.8 ms]
  Range (min … max):    18.4 ms …  20.1 ms    159 runs
 
Benchmark 2: /Users/shunsukesuzuki/.local/share/aquaproj-aqua/pkgs/github_release/github.com/suzuki-shunsuke/tfcmt/v3.2.1/tfcmt_darwin_arm64.tar.gz/tfcmt -v
  Time (mean ± σ):       3.2 ms ±   0.1 ms    [User: 1.9 ms, System: 1.0 ms]
  Range (min … max):     3.1 ms …   3.8 ms    898 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet PC without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
Summary
  '/Users/shunsukesuzuki/.local/share/aquaproj-aqua/pkgs/github_release/github.com/suzuki-shunsuke/tfcmt/v3.2.1/tfcmt_darwin_arm64.tar.gz/tfcmt -v' ran
    5.92 ± 0.16 times faster than 'tfcmt -v'
```

```console
$ hyperfine -N --warmup 3 'tfcmt -v' '/Users/shunsukesuzuki/.local/share/aquaproj-aqua/pkgs/github_release/github.com/suzuki-shunsuke/tfcmt/v3.2.1/tfcmt_darwin_arm64.tar.gz/tfcmt -v'
Benchmark 1: tfcmt -v
  Time (mean ± σ):      15.0 ms ±   0.2 ms    [User: 9.9 ms, System: 5.3 ms]
  Range (min … max):    14.4 ms …  15.5 ms    198 runs
 
Benchmark 2: /Users/shunsukesuzuki/.local/share/aquaproj-aqua/pkgs/github_release/github.com/suzuki-shunsuke/tfcmt/v3.2.1/tfcmt_darwin_arm64.tar.gz/tfcmt -v
  Time (mean ± σ):       3.3 ms ±   0.1 ms    [User: 1.9 ms, System: 1.0 ms]
  Range (min … max):     3.1 ms …   4.0 ms    928 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet PC without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
Summary
  '/Users/shunsukesuzuki/.local/share/aquaproj-aqua/pkgs/github_release/github.com/suzuki-shunsuke/tfcmt/v3.2.1/tfcmt_darwin_arm64.tar.gz/tfcmt -v' ran
    4.62 ± 0.13 times faster than 'tfcmt -v'
```

```console
$ hyperfine -N --warmup 3 'env AQUA_CONFIG=aqua.yaml tfcmt -v' 'env AQUA_CONFIG=aqua-json.yaml tfcmt -v'
Benchmark 1: env AQUA_CONFIG=aqua.yaml tfcmt -v
  Time (mean ± σ):      24.4 ms ±   0.4 ms    [User: 17.1 ms, System: 7.1 ms]
  Range (min … max):    23.6 ms …  25.4 ms    121 runs
 
Benchmark 2: env AQUA_CONFIG=aqua-json.yaml tfcmt -v
  Time (mean ± σ):      20.2 ms ±   0.4 ms    [User: 11.4 ms, System: 6.5 ms]
  Range (min … max):    19.4 ms …  21.1 ms    144 runs
 
Summary
  'env AQUA_CONFIG=aqua-json.yaml tfcmt -v' ran
    1.21 ± 0.03 times faster than 'env AQUA_CONFIG=aqua.yaml tfcmt -v'
```