@suzuki-shunsuke 2022-05-01
I tried goccy/go-json, but there was no change compared with encoding/json.

```console
$ hyperfine -N --warmup 3 'env AQUA_CONFIG=aqua.yaml tfcmt -v' 'env AQUA_CONFIG=aqua-json.yaml tfcmt -v'
Benchmark 1: env AQUA_CONFIG=aqua.yaml tfcmt -v
  Time (mean ± σ):      24.6 ms ±   0.5 ms    [User: 16.9 ms, System: 7.2 ms]
  Range (min … max):    23.2 ms …  26.5 ms    120 runs
 
Benchmark 2: env AQUA_CONFIG=aqua-json.yaml tfcmt -v
  Time (mean ± σ):      19.9 ms ±   0.4 ms    [User: 12.2 ms, System: 7.0 ms]
  Range (min … max):    18.7 ms …  21.3 ms    152 runs
 
Summary
  'env AQUA_CONFIG=aqua-json.yaml tfcmt -v' ran
    1.24 ± 0.04 times faster than 'env AQUA_CONFIG=aqua.yaml tfcmt -v'
```